Add files via upload
I added my changes which iterates over the entire line between the two points given. My changes can be found at the bottom of the collision check. It works with path4 but I could not get it to work with the other paths which is why I am commiting it a separate branch. If you run it with the other paths such as path2 you could be waiting a very very very long time.